Refactor: 컨트롤러를 기능별 책임에 따라 분리
- 단일 책임 원칙을 준수하고 컨트롤러의 가독성 및 유지보수성을 향상시키기 위해 기존 RentalController를 기능별 책임에 따라 세분화했습니다.
